requirements.txt
Flask==1.0.2
gunicorn==19.5.0
gevent==1.4.0
sqlalchemy==1.3.13
Flask-JWT
uwsgi
mysqlclient
python-dateutil
pycrypto
google-auth
flask_cors